// apps/kyc/src/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const KycTier = {
  UNVERIFIED: 'UNVERIFIED',
  BASIC: 'BASIC',
  VERIFIED: 'VERIFIED'
} as const

export type KycTier = (typeof KycTier)[keyof typeof KycTier]


export const KycStatus = {
  NONE: 'NONE',
  PENDING_REVIEW: 'PENDING_REVIEW',
  APPROVED: 'APPROVED',
  REJECTED: 'REJECTED'
} as const

export type KycStatus = (typeof KycStatus)[keyof typeof KycStatus]


export const DocType = {
  PASSPORT: 'PASSPORT',
  DRIVERS_LICENSE: 'DRIVERS_LICENSE',
  SELFIE: 'SELFIE',
  UTILITY_BILL: 'UTILITY_BILL',
  OTHER: 'OTHER'
} as const

export type DocType = (typeof DocType)[keyof typeof DocType]